requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
asyncpg==0.29.0
certifi==2025.6.15
click==8.2.1
exceptiongroup==1.3.0
fastapi==0.104.1
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pika==1.3.2
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic_core==2.14.1
PyPDF2==3.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
typing_extensions==4.14.0
uvicorn==0.24.0
pytest-asyncio==0.23.6
